refactor(node-primitives): one copy of the node and bundle fixtures
The acquisition test mirrored `tests/common/mod.rs::node_client()` and carried
its own signed-bundle packer, which is what five dev-dependencies on
calimero-context-client were paying for. Both now come from a `testing` feature
of calimero-node-primitives, the shape `calimero-storage` already uses, and the
crate's own tests take it from a dev-dependency on itself so they cannot
silently lose the fixtures if the other consumer goes away.

The bundle's package and version are no longer re-asserted from the installed
metadata: that is the installer's claim and `bundle_installation` makes it
field for field.
