+++ crates/context/primitives/Cargo.toml
@@ -35,10 +35,6 @@ calimero-utils-actix.workspace = true
 [dev-dependencies]
 calimero-blobstore.workspace = true
 calimero-network-primitives.workspace = true
-camino.workspace = true
-ed25519-dalek.workspace = true
-flate2.workspace = true
-serde_json.workspace = true
-tar.workspace = true
+calimero-node-primitives = { workspace = true, features = ["testing"] }
 tempfile = "3.23"
 tokio = { workspace = true, features = ["macros", "rt-multi-thread", "sync"] }
