+++ crates/node/primitives/src/lib.rs
@@ -12,6 +12,8 @@ pub use join_bundle::JoinBundle;
 pub mod messages;
 pub mod sync;
 pub mod sync_status;
+#[cfg(any(test, feature = "testing"))]
+pub mod test_fixtures;
 pub use sync_status::{SyncState, SyncStatusSnapshot};
 pub mod topic_manager;
 pub use topic_manager::TopicManager;
